requirements.txt
Django>=4.2
python-dotenv>=1.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-cors-headers>=4.3
gunicorn>=21.0
psycopg2-binary>=2.9
Pillow>=10.0
